apps/music_scene/requirements.txt
python-fasthtml==0.3.6
sqlite-utils==3.37